+++ src/index.ts
@@ -3,5 +3,7 @@ export * from "./engine/envParser";
 export * from "./engine/validator";
 export * from "./engine/interactive";
 export * from "./engine/envWriter";
+export * from "./engine/envGenerator";
+export * from "./engine/codeScanner";
 export * from "./config/loadConfig";
 export * from "./types";